+++ player.cc
@@ -9,7 +9,7 @@
 #include "character.h"
 #include "player.h"
 
-player::player(coord pos, int hp, int atk, int def, std::string name):character{pos,hp,atk,def,name}
+player::player(coord pos, int hp, double atk, double def, std::string name):character{pos,hp,atk,def,name}
 {
 	maxHP = hp;
 }
